The modifications remove a bug which occurs when dealing with special numbers and the callerid.conf file.  Regular numbers such as +4940555 can be noted in the following way in a NUMBER line: NUMBER=+49-40-555, +49-69-555
For special numbers prior to this modifications there are strong
unmentioned limitations.  Only one number per NUMBER line is allowed
without any of the options shown above.
Allowed: NUMBER=01802555
Not allowed/not working: NUMBER=01802-555
also not working: NUMBER=0180255, 0700555

This modification brings the later two examples in function.  About the
background of the problem:  isdnlog/isdnrep and other know a sort of
telephone numbers called `special numbers', which are noted with S: and
N: Tags in the rate-file. These numbers are per definition only available
within the country.  In order to emphasize this and for other more
practical reasons, the normalized format for this numbers is not the
usual +<countrycode><areacode><local number> but the national format
like 01802555.

While the logged numbers in /var/log/isdn.log or so follow this convention,
such entries in callerid.conf got internally converted to the normalized
format (+49180255 in case of the last mentioned number) so that the
comparison executed by isdnlog or isdnrep did not show the expected
result.

The modification activates the advanced feature in the notation of normal
numbers for special numbers.

IMPORTANT NOTE:  As the resulting library is located outside the isdnlog
subdirectory, this changes may influence other parts of the isdn4k-utils.